requirements.txt
flask==1.1.2
flask-bootstrap==3.3.7.1
gunicorn==20.0.4
python-dotenv==0.14.0
numpy==1.18.0
pypng==0.0.20
